Moving CollapsibleToolbar state into separate file.
